Fixed the join command, TA
The join_cmd() used to receive, more than often, a string full of garbage as its password field but it has been fixed. Now this command should be fully functional terminal wise. For KVIrc is another story, as it puts the name of the channel in its users list, no idea why. Also there is a 50/50 chance that it will visually create the channel as does basic KVIrc, or won't. It should be a desync issue.